+++ changelog.md
@@ -6,6 +6,15 @@
 
 [commit](https://github.com/sbrstrkkdwmdr/sbrbot)</br>
 
+### Fixed
+
+- fix lazer scores with classic mod showing 0 total score
+
+### Added
+
+- scoreparse shows total score
+- scoreparse shows country rank (if available)
+
 ## [4.1.2] - 2025-02-11
 
 [commit](https://github.com/sbrstrkkdwmdr/sbrbot/commit/bb26b635628a8aadde42b67c7e9f3f4194bcfc96)</br>